+++ src/Captcha.js
@@ -14,7 +14,7 @@ class Captcha {
 
         let form = $(this.html).find("*[id^=formVerify]");
         this.formAction = form.attr('action');
-        // this.url = Slickiss.absolute(form.find("input[name='reUrl']").val());
+        // this.url = S.absolute(form.find("input[name='reUrl']").val());
         this.imgs = form.find('img');
         this.q1 = form.find('span').eq(0).text().trim();
         this.q2 = form.find('span').eq(1).text().trim();